==> NIR v0.1/0. Data comb prop NIR.ipynb <==
import pandas as pd
from sklearn.cluster import KMeans
import matplotlib.pyplot as plt
import seaborn as sns

# Load the data
country_mapping = pd.read_excel(r"C:\Users\pingk\Downloads\fadhli nitip\country_mapping.xlsx")
country_2digit = pd.read_excel(r"C:\Users\pingk\Downloads\fadhli nitip\country_2digit.xlsx")
provincial_code = pd.read_excel(r"C:\Users\pingk\Downloads\fadhli nitip\provincial code.xlsx")
three_group_precipitation = pd.read_excel(r"C:\Users\pingk\Downloads\fadhli nitip\three_group_precipitation.xlsx")
four_group_precipitation = pd.read_excel(r"C:\Users\pingk\Downloads\fadhli nitip\four_group_precipitation.xlsx")
island_mapping = pd.read_excel(r"C:\Users\pingk\Downloads\fadhli nitip\island code.xlsx")
thai_mapping = pd.read_excel(r"C:\Users\pingk\Downloads\fadhli nitip\thai_vs_nothai.xlsx")

data_rt4 = pd.read_csv(r"C:\Users\pingk\Downloads\fadhli nitip\fadhli nitip\csv\data_NIR_1.csv")

# Inspect the first few rows of each dataset
print("Raw Data:")
print(data_rt4.head())
print(country_mapping.head())
print(country_2digit.head())
print(provincial_code.head())
print(three_group_precipitation.head())
print(four_group_precipitation.head())
print(island_mapping.head())
print(thai_mapping.head())

# Merge data_rt4 with provincial_code on 'prov_char'
data_rt4_with_prov_code = pd.merge(data_rt4, provincial_code, how='left', left_on='prov_char', right_on='prov_char')

# Merge the result with three_group_precipitation
data_rt4_with_3precip = pd.merge(data_rt4_with_prov_code, three_group_precipitation, how='left', left_on='prov_char', right_on='prov_char')

# Merge the result with four_group_precipitation
data_rt4_with_3p4p = pd.merge(data_rt4_with_3precip, four_group_precipitation, how='left', left_on='prov_char', right_on='prov_char')

# Merge the result with country_mapping
data_rt4_with_countryname = pd.merge(data_rt4_with_3p4p, country_mapping, how='left', left_on='country_code', right_on='country_code')

# Merge the result with thai_mapping
data_rt4_with_thai = pd.merge(data_rt4_with_countryname, thai_mapping, how='left', left_on='country_char', right_on='country_char')

# Merge the result with island_mapping
data_rt4_with_sixcat = pd.merge(data_rt4_with_thai, island_mapping, how='left', left_on='prov_char', right_on='prov_char')

# Inspect the first few rows of each dataset
print("Data with Additional Groupings:")
print(data_rt4_with_sixcat.head())

# Input your dataset here

asik  = data_rt4_with_sixcat.drop(
        columns=[
            col for col in data_rt4_with_sixcat.columns 
            if col in ['country_code', 'prov_code', 'three_group_precip', 'four_group_precip', 'th_vs_noth', 'island_code'] 
            #or (col.replace('.', '', 1).isdigit() and any(float(col) >= start and float(col) < end for start, end in [(3100, 4001), (1870, 2700)]))
            #or (col.replace('.', '', 1).isdigit() and 822.01 <= float(col) < 879.99)
        ]
    )

asik.to_csv('C:/Users/pingk/Downloads/fadhli nitip/asik_NIR_1.csv', index=False)
